+++ js/util.js
@@ -1,0 +1,19 @@
+function getRandomNumber (min, max) {
+  if (min >= 0 && max > min) {
+    return Math.round(Math.random() * (max - min)) + min;
+  }
+
+  return 'Что-то пошло не так';
+}
+
+function getCheckString (string, maxLength) {
+  if (string.length <= maxLength) {
+    return true;
+  }
+
+  return false;
+}
+
+getCheckString('Привет, кекс', 140);
+
+export {getRandomNumber};
